move `Context` back to `core` module
- Move `Cache` into the new module, since it's referenced from
  `Schema`.  The cache implementations will also need to be moved.

- Move `Context.scala` back into the `core` module.  There's too much
  that the `Context` class pulls in.  This will need a trait that
  `Context` will implement.  Instead, extract the `Action` hierarchy,
  `MappingDeferred`, and `case class Args` from the file into their own
  files (leaving `object Args`).

- Move `Deferred` and `FieldTag` (and `SubscriptionField`) into the new
  module, into their own files.

- Move `SchemaComparator.scala` back into the `core` module.  We'll
  remove the `extend` and `compare` methods from `Schema` instead.
