+++ modules/core/src/main/scala/sangria/execution/middleware.scala
@@ -158,15 +158,11 @@ object BeforeFieldResult {
     BeforeFieldResult[Ctx, Unit]((), attachment = Some(a))
 }
 
-trait FieldTag
-
 case class StringTag(name: String)
 
 object FieldTag {
   implicit def stringTag(s: String): StringTag = StringTag(s)
   implicit def symbolTag(s: Symbol): StringTag = StringTag(s.name)
 }
-
-case class SubscriptionField[S[_]](stream: SubscriptionStream[S]) extends FieldTag
 
 case class Extension[In](data: In)(implicit val iu: InputUnmarshaller[In])
