+++ modules/core/src/main/scala/sangria/schema/Context.scala
@@ -5,173 +5,9 @@ import sangria.execution._
 import sangria.marshalling._
 import sangria.ast.SourceMapper
 import sangria.{ast, introspection}
-import sangria.execution.deferred.Deferred
-import sangria.streaming.SubscriptionStream
 import sangria.util.Cache
 
-import scala.concurrent.{ExecutionContext, Future}
 import scala.reflect.ClassTag
-import scala.util.{Failure, Try}
-import scala.util.control.NonFatal
-
-sealed trait Action[+Ctx, +Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): Action[Ctx, NewVal]
-}
-sealed trait LeafAction[+Ctx, +Val] extends Action[Ctx, Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): LeafAction[Ctx, NewVal]
-}
-sealed trait ReduceAction[+Ctx, +Val] extends Action[Ctx, Val] {
-  def map[NewVal](fn: Val => NewVal)(implicit ec: ExecutionContext): LeafAction[Ctx, NewVal]
-}
-
-object ReduceAction {
-  implicit def futureAction[Ctx, Val](value: Future[Val]): ReduceAction[Ctx, Val] = FutureValue(
-    value)
-  implicit def tryAction[Ctx, Val](value: Try[Val]): ReduceAction[Ctx, Val] = TryValue(value)
-  implicit def defaultAction[Ctx, Val](value: Val): ReduceAction[Ctx, Val] = Value(value)
-}
-
-object Action extends LowPrioActions {
-  def sequence[Ctx, Val](actions: Seq[LeafAction[Ctx, Val]]): SequenceLeafAction[Ctx, Val] =
-    SequenceLeafAction[Ctx, Val](actions)
-
-  def apply[Ctx, Val](a: Action[Ctx, Val]): Action[Ctx, Val] = a
-
-  implicit def deferredAction[Ctx, Val](value: Deferred[Val]): LeafAction[Ctx, Val] = DeferredValue(
-    value)
-  implicit def tryAction[Ctx, Val](value: Try[Val]): LeafAction[Ctx, Val] = TryValue(value)
-}
-
-trait LowPrioActions extends LowestPrioActions {
-  implicit def deferredFutureAction[Ctx, Val](value: Future[Deferred[Val]]): LeafAction[Ctx, Val] =
-    DeferredFutureValue(value)
-}
-
-trait LowestPrioActions {
-  implicit def futureAction[Ctx, Val](value: Future[Val]): LeafAction[Ctx, Val] = FutureValue(value)
-  implicit def defaultAction[Ctx, Val](value: Val): LeafAction[Ctx, Val] = Value(value)
-}
-
-object LeafAction {
-  def sequence[Ctx, Val](actions: Seq[LeafAction[Ctx, Val]]): SequenceLeafAction[Ctx, Val] =
-    SequenceLeafAction[Ctx, Val](actions)
-
-  def apply[Ctx, Val](a: LeafAction[Ctx, Val]): LeafAction[Ctx, Val] = a
-}
-
-case class Value[Ctx, Val](value: Val) extends LeafAction[Ctx, Val] with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): LeafAction[Ctx, NewVal] =
-    try Value(fn(value))
-    catch {
-      case NonFatal(e) => TryValue(Failure(e))
-    }
-}
-
-case class TryValue[Ctx, Val](value: Try[Val])
-    extends LeafAction[Ctx, Val]
-    with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): TryValue[Ctx, NewVal] =
-    TryValue(value.map(fn))
-}
-
-case class PartialValue[Ctx, Val](value: Val, errors: Vector[Throwable])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): LeafAction[Ctx, NewVal] =
-    try PartialValue(fn(value), errors)
-    catch {
-      case NonFatal(e) => TryValue(Failure(e))
-    }
-}
-
-case class FutureValue[Ctx, Val](value: Future[Val])
-    extends LeafAction[Ctx, Val]
-    with ReduceAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): FutureValue[Ctx, NewVal] =
-    FutureValue(value.map(fn))
-}
-
-case class PartialFutureValue[Ctx, Val](value: Future[PartialValue[Ctx, Val]])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): PartialFutureValue[Ctx, NewVal] =
-    PartialFutureValue(value.map(_.map(fn) match {
-      case v: PartialValue[Ctx, NewVal] => v
-      case TryValue(Failure(e)) => throw e
-      case v => throw new IllegalStateException("Unexpected result from `PartialValue.map`: " + v)
-    }))
-}
-
-case class DeferredValue[Ctx, Val](value: Deferred[Val]) extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): DeferredValue[Ctx, NewVal] =
-    DeferredValue(MappingDeferred(value, (v: Val) => (fn(v), Vector.empty)))
-
-  def mapWithErrors[NewVal](fn: Val => (NewVal, Vector[Throwable])): DeferredValue[Ctx, NewVal] =
-    DeferredValue(MappingDeferred(value, fn))
-}
-
-case class DeferredFutureValue[Ctx, Val](value: Future[Deferred[Val]])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): DeferredFutureValue[Ctx, NewVal] =
-    DeferredFutureValue(value.map(MappingDeferred(_, (v: Val) => (fn(v), Vector.empty))))
-
-  def mapWithErrors[NewVal](fn: Val => (NewVal, Vector[Throwable]))(implicit
-      ec: ExecutionContext): DeferredFutureValue[Ctx, NewVal] =
-    DeferredFutureValue(value.map(MappingDeferred(_, fn)))
-}
-
-case class SequenceLeafAction[Ctx, Val](value: Seq[LeafAction[Ctx, Val]])
-    extends LeafAction[Ctx, Seq[Val]] {
-  override def map[NewVal](fn: Seq[Val] => NewVal)(implicit
-      ec: ExecutionContext): MappedSequenceLeafAction[Ctx, Val, NewVal] =
-    new MappedSequenceLeafAction[Ctx, Val, NewVal](this, fn)
-}
-
-class MappedSequenceLeafAction[Ctx, Val, NewVal](
-    val action: SequenceLeafAction[Ctx, Val],
-    val mapFn: Seq[Val] => NewVal)
-    extends LeafAction[Ctx, NewVal] {
-  override def map[NewNewVal](fn: NewVal => NewNewVal)(implicit
-      ec: ExecutionContext): MappedSequenceLeafAction[Ctx, Val, NewNewVal] =
-    new MappedSequenceLeafAction[Ctx, Val, NewNewVal](action, v => fn(mapFn(v)))
-}
-
-class UpdateCtx[Ctx, Val](val action: LeafAction[Ctx, Val], val nextCtx: Val => Ctx)
-    extends Action[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): MappedUpdateCtx[Ctx, Val, NewVal] =
-    new MappedUpdateCtx[Ctx, Val, NewVal](action, nextCtx, fn)
-}
-
-class MappedUpdateCtx[Ctx, Val, NewVal](
-    val action: LeafAction[Ctx, Val],
-    val nextCtx: Val => Ctx,
-    val mapFn: Val => NewVal)
-    extends Action[Ctx, NewVal] {
-  override def map[NewNewVal](fn: NewVal => NewNewVal)(implicit
-      ec: ExecutionContext): MappedUpdateCtx[Ctx, Val, NewNewVal] =
-    new MappedUpdateCtx[Ctx, Val, NewNewVal](action, nextCtx, v => fn(mapFn(v)))
-}
-
-object UpdateCtx {
-  def apply[Ctx, Val](action: LeafAction[Ctx, Val])(newCtx: Val => Ctx): UpdateCtx[Ctx, Val] =
-    new UpdateCtx(action, newCtx)
-}
-
-private[sangria] case class SubscriptionValue[Ctx, Val, S[_]](
-    source: Val,
-    stream: SubscriptionStream[S])
-    extends LeafAction[Ctx, Val] {
-  override def map[NewVal](fn: Val => NewVal)(implicit
-      ec: ExecutionContext): SubscriptionValue[Ctx, NewVal, S] =
-    throw new IllegalStateException(
-      "`map` is not supported subscription actions. Action is only intended for internal use.")
-}
 
 case class ProjectionName(name: String) extends FieldTag
 case object ProjectionExclude extends FieldTag
@@ -213,9 +49,6 @@ case class ProjectedName(
     loop(this)
   }
 }
-
-case class MappingDeferred[A, +B](deferred: Deferred[A], mapFn: A => (B, Vector[Throwable]))
-    extends Deferred[B]
 
 trait WithArguments {
   def args: Args
@@ -439,102 +272,6 @@ case class Context[Ctx, Val](
       case a if clazz.isAssignableFrom(a.getClass) => a.asInstanceOf[T]
     }
   }
-}
-
-case class Args(
-    raw: Map[String, Any],
-    argsWithDefault: Set[String],
-    optionalArgs: Set[String],
-    undefinedArgs: Set[String],
-    defaultInfo: Cache[String, Any]) {
-  private def getAsOptional[T](name: String): Option[T] =
-    raw.get(name).asInstanceOf[Option[Option[T]]].flatten
-
-  private def invariantExplicitlyNull(name: String) =
-    throw new IllegalArgumentException(
-      s"Optional argument '$name' accessed as a non-optional argument (it has a default value), but query explicitly set argument to `null`.")
-
-  private def invariantNotProvided(name: String) =
-    throw new IllegalArgumentException(
-      s"Optional argument '$name' accessed as a non-optional argument, but it was not provided in the query and argument does not define a default value.")
-
-  def arg[T](arg: Argument[T]): T =
-    if (optionalArgs.contains(arg.name) && argsWithDefault.contains(arg.name) && defaultInfo
-        .contains(arg.name))
-      getAsOptional[T](arg.name).getOrElse(defaultInfo(arg.name).asInstanceOf[T])
-    else if (optionalArgs.contains(arg.name) && argsWithDefault.contains(arg.name))
-      getAsOptional[T](arg.name).getOrElse(invariantExplicitlyNull(arg.name))
-    else if (optionalArgs.contains(arg.name))
-      getAsOptional[Any](arg.name).asInstanceOf[T]
-    else
-      raw(arg.name).asInstanceOf[T]
-
-  def arg[T](name: String): T =
-    if (optionalArgs.contains(name) && argsWithDefault.contains(name) && defaultInfo.contains(name))
-      getAsOptional[T](name).getOrElse(defaultInfo(name).asInstanceOf[T])
-    else if (optionalArgs.contains(name) && argsWithDefault.contains(name))
-      getAsOptional[T](name).getOrElse(invariantExplicitlyNull(name))
-    else if (optionalArgs.contains(name))
-      getAsOptional[T](name).getOrElse(invariantNotProvided(name))
-    else
-      raw(name).asInstanceOf[T]
-
-  def argOpt[T](name: String): Option[T] = getAsOptional(name)
-
-  def argOpt[T](arg: Argument[T]): Option[T] =
-    if (optionalArgs.contains(arg.name))
-      getAsOptional[T](arg.name)
-    else
-      raw.get(arg.name).asInstanceOf[Option[T]]
-
-  def argDefinedInQuery(name: String): Boolean = !undefinedArgs.contains(name)
-  def argDefinedInQuery(arg: Argument[_]): Boolean = argDefinedInQuery(arg.name)
-
-  def withArgs[A1, R](arg1: Argument[A1])(fn: A1 => R): R = fn(arg(arg1))
-  def withArgs[A1, A2, R](arg1: Argument[A1], arg2: Argument[A2])(fn: (A1, A2) => R): R =
-    fn(arg(arg1), arg(arg2))
-  def withArgs[A1, A2, A3, R](arg1: Argument[A1], arg2: Argument[A2], arg3: Argument[A3])(
-      fn: (A1, A2, A3) => R): R = fn(arg(arg1), arg(arg2), arg(arg3))
-  def withArgs[A1, A2, A3, A4, R](
-      arg1: Argument[A1],
-      arg2: Argument[A2],
-      arg3: Argument[A3],
-      arg4: Argument[A4])(fn: (A1, A2, A3, A4) => R): R =
-    fn(arg(arg1), arg(arg2), arg(arg3), arg(arg4))
-  def withArgs[A1, A2, A3, A4, A5, R](
-      arg1: Argument[A1],
-      arg2: Argument[A2],
-      arg3: Argument[A3],
-      arg4: Argument[A4],
-      arg5: Argument[A5])(fn: (A1, A2, A3, A4, A5) => R): R =
-    fn(arg(arg1), arg(arg2), arg(arg3), arg(arg4), arg(arg5))
-  def withArgs[A1, A2, A3, A4, A5, A6, R](
-      arg1: Argument[A1],
-      arg2: Argument[A2],
-      arg3: Argument[A3],
-      arg4: Argument[A4],
-      arg5: Argument[A5],
-      arg6: Argument[A6])(fn: (A1, A2, A3, A4, A5, A6) => R): R =
-    fn(arg(arg1), arg(arg2), arg(arg3), arg(arg4), arg(arg5), arg(arg6))
-  def withArgs[A1, A2, A3, A4, A5, A6, A7, R](
-      arg1: Argument[A1],
-      arg2: Argument[A2],
-      arg3: Argument[A3],
-      arg4: Argument[A4],
-      arg5: Argument[A5],
-      arg6: Argument[A6],
-      arg7: Argument[A7])(fn: (A1, A2, A3, A4, A5, A6, A7) => R): R =
-    fn(arg(arg1), arg(arg2), arg(arg3), arg(arg4), arg(arg5), arg(arg6), arg(arg7))
-  def withArgs[A1, A2, A3, A4, A5, A6, A7, A8, R](
-      arg1: Argument[A1],
-      arg2: Argument[A2],
-      arg3: Argument[A3],
-      arg4: Argument[A4],
-      arg5: Argument[A5],
-      arg6: Argument[A6],
-      arg7: Argument[A7],
-      arg8: Argument[A8])(fn: (A1, A2, A3, A4, A5, A6, A7, A8) => R): R =
-    fn(arg(arg1), arg(arg2), arg(arg3), arg(arg4), arg(arg5), arg(arg6), arg(arg7), arg(arg8))
 }
 
 object Args {
