+++ modules/schema/src/main/scala/sangria/execution/FieldTag.scala
@@ -1,0 +1,7 @@
+package sangria.execution
+
+import sangria.streaming.SubscriptionStream
+
+trait FieldTag
+
+case class SubscriptionField[S[_]](stream: SubscriptionStream[S]) extends FieldTag
