+++ modules/schema/src/main/scala/sangria/execution/deferred/Deferred.scala
@@ -1,0 +1,3 @@
+package sangria.execution.deferred
+
+trait Deferred[+T]
